[Rgen] Fix an error from PR #23115
The refactor was wrong and we refactored the
GetTrampolineInitializationByRefArgument whichs does native -> managed
instead of GetTrampolineNativeInvokeArgumentInitializations.

This commit undoos that mistake.
